Remove `ChannelMutableState` interface
