+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/StateRegistry.kt
@@ -25,7 +25,6 @@ import io.getstream.chat.android.client.persistance.repository.MessageRepository
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.offline.plugin.state.channel.ChannelState
 import io.getstream.chat.android.offline.plugin.state.channel.internal.ChannelMutableState
-import io.getstream.chat.android.offline.plugin.state.channel.internal.ChannelMutableStateImpl
 import io.getstream.chat.android.offline.plugin.state.channel.thread.ThreadState
 import io.getstream.chat.android.offline.plugin.state.channel.thread.internal.ThreadMutableState
 import io.getstream.chat.android.offline.plugin.state.channel.thread.internal.ThreadMutableStateImpl
@@ -83,7 +82,7 @@ public class StateRegistry private constructor(
      */
     public fun channel(channelType: String, channelId: String): ChannelState {
         return channels.getOrPut(channelType to channelId) {
-            ChannelMutableStateImpl(channelType, channelId, scope, userStateFlow, latestUsers)
+            ChannelMutableState(channelType, channelId, scope, userStateFlow, latestUsers)
         }
     }
