+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/offline/plugin/state/channel/internal/ChannelMutableState.kt
@@ -17,6 +17,9 @@
 package io.getstream.chat.android.offline.plugin.state.channel.internal
 
 import io.getstream.chat.android.client.events.TypingStartEvent
+import io.getstream.chat.android.client.extensions.internal.updateUsers
+import io.getstream.chat.android.client.extensions.internal.wasCreatedAfter
+import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.ChannelUserRead
 import io.getstream.chat.android.client.models.Config
 import io.getstream.chat.android.client.models.Member
@@ -25,148 +28,344 @@ import io.getstream.chat.android.client.models.TypingEvent
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.offline.model.channel.ChannelData
 import io.getstream.chat.android.offline.plugin.state.channel.ChannelState
+import io.getstream.chat.android.offline.plugin.state.channel.MessagesState
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.filterNotNull
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
 import java.util.Date
 
-@Suppress("VariableNaming", "TooManyFunctions")
 /** State container with mutable data of a channel.*/
-internal interface ChannelMutableState : ChannelState {
+internal class ChannelMutableState(
+    override val channelType: String,
+    override val channelId: String,
+    private val scope: CoroutineScope,
+    private val userFlow: StateFlow<User?>,
+    latestUsers: StateFlow<Map<String, User>>,
+) : ChannelState {
 
-    /** Sorted version of messages. */
-    val sortedMessages: StateFlow<List<Message>>
+    override val cid: String = "%s:%s".format(channelType, channelId)
 
-    /** The message collection of this channel. */
-    val messageList: StateFlow<List<Message>>
+    private val _messages = MutableStateFlow<Map<String, Message>>(emptyMap())
+    private val _typing = MutableStateFlow(TypingEvent(channelId, emptyList()))
+    private val _typingChatEvents = MutableStateFlow<Map<String, TypingStartEvent>>(emptyMap())
+    private val _rawReads = MutableStateFlow<Map<String, ChannelUserRead>>(emptyMap())
+    private val _members = MutableStateFlow<Map<String, Member>>(emptyMap())
+    private val _oldMessages = MutableStateFlow<Map<String, Message>>(emptyMap())
+    private val _watchers = MutableStateFlow<Map<String, User>>(emptyMap())
+    private val _watcherCount = MutableStateFlow(0)
+    private val _read = MutableStateFlow<ChannelUserRead?>(null)
+    private val _endOfNewerMessages = MutableStateFlow(false)
+    private val _endOfOlderMessages = MutableStateFlow(false)
+    private val _loading = MutableStateFlow(false)
+    private val _hidden = MutableStateFlow(false)
+    private val _muted = MutableStateFlow(false)
+    private val _channelData = MutableStateFlow<ChannelData?>(null)
+    private val _repliedMessage = MutableStateFlow<Message?>(null)
+    private val _unreadCount = MutableStateFlow(0)
+    private val _membersCount = MutableStateFlow(0)
+    private val _insideSearch = MutableStateFlow(false)
+    private val _loadingOlderMessages = MutableStateFlow(false)
+    private val _loadingNewerMessages = MutableStateFlow(false)
+    private val _lastMessageAt = MutableStateFlow<Date?>(null)
 
     /** raw version of messages. */
     var rawMessages: Map<String, Message>
+        get() = _messages.value
+        set(value) { _messages.value = value }
 
     /** raw version of reads. */
     var rawReads: Map<String, ChannelUserRead>
+        get() = _rawReads.value
+        set(value) { _rawReads.value = value }
 
-    /** raw version of reads. */
+    /** raw version of members. */
     var rawMembers: Map<String, Member>
+        get() = _members.value
+        set(value) { _members.value = value }
 
     /** raw version of old messages. */
     var rawOldMessages: Map<String, Message>
+        get() = _oldMessages.value
+        set(value) { _oldMessages.value = value }
 
-    /** raw version of old watchers. */
+    /** raw version of watchers. */
     var rawWatchers: Map<String, User>
+        get() = _watchers.value
+        set(value) { _watchers.value = value }
 
     /** raw version of typing. */
     var rawTyping: Map<String, TypingStartEvent>
+        get() = _typingChatEvents.value
+        set(value) { _typingChatEvents.value = value }
 
     /** the date of the last message */
     var lastMessageAt: Date?
+        get() = _lastMessageAt.value
+        set(value) { _lastMessageAt.value = value }
+
+    /** Channel config data. */
+    private val _channelConfig: MutableStateFlow<Config> = MutableStateFlow(Config())
 
     /** the data to hide messages before */
-    var hideMessagesBefore: Date?
+    var hideMessagesBefore: Date? = null
+
+    /** The raw message list updated by recent users value. */
+    val messageList: StateFlow<List<Message>> =
+        _messages.combine(latestUsers) { messageMap, userMap -> messageMap.values.updateUsers(userMap) }
+            .stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    /** a list of messages sorted by message.createdAt */
+    private val sortedVisibleMessages: StateFlow<List<Message>> =
+        messagesTransformation(messageList).stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    private val _messagesState: StateFlow<MessagesState> =
+        _loading.combine(sortedVisibleMessages) { loading: Boolean, messages: List<Message> ->
+            when {
+                loading -> MessagesState.Loading
+                messages.isEmpty() -> MessagesState.OfflineNoResults
+                else -> MessagesState.Result(messages)
+            }
+        }.stateIn(scope, SharingStarted.Eagerly, MessagesState.NoQueryActive)
+
+    private fun messagesTransformation(messages: Flow<Collection<Message>>): StateFlow<List<Message>> {
+        return messages.combine(userFlow) { messageCollection, user ->
+            messageCollection.asSequence()
+                .filter { it.parentId == null || it.showInChannel }
+                .filter { it.user.id == user?.id || !it.shadowed }
+                .filter { hideMessagesBefore == null || it.wasCreatedAfter(hideMessagesBefore) }
+                .sortedBy { it.createdAt ?: it.createdLocallyAt }
+                .toList()
+        }.stateIn(scope, SharingStarted.Eagerly, emptyList())
+    }
+
+    internal var lastMarkReadEvent: Date? = null
+    internal var lastKeystrokeAt: Date? = null
 
     /** The date of the last typing event. */
-    var lastStartTypingEvent: Date?
+    var lastStartTypingEvent: Date? = null
+    internal var keystrokeParentMessageId: String? = null
+
+    /** Sorted version of messages. */
+    val sortedMessages: StateFlow<List<Message>> = messageList.map {
+        it.sortedBy { message -> message.createdAt ?: message.createdLocallyAt }
+            .filter { message -> hideMessagesBefore == null || message.wasCreatedAfter(hideMessagesBefore) }
+    }.stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    override val repliedMessage: StateFlow<Message?> = _repliedMessage
+
+    /** Channel config data */
+    override val channelConfig: StateFlow<Config> = _channelConfig
+
+    override val messages: StateFlow<List<Message>> = sortedVisibleMessages
+
+    override val messagesState: StateFlow<MessagesState> = _messagesState
+    override val oldMessages: StateFlow<List<Message>> = messagesTransformation(_oldMessages.map { it.values })
+    override val watcherCount: StateFlow<Int> = _watcherCount
+    override val watchers: StateFlow<List<User>> =
+        _watchers.combine(latestUsers) { watcherMap, userMap -> watcherMap.values.updateUsers(userMap) }
+            .map { it.sortedBy(User::createdAt) }
+            .stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    override val typing: StateFlow<TypingEvent> = _typing
+
+    override val reads: StateFlow<List<ChannelUserRead>> = _rawReads
+        .map { it.values.sortedBy(ChannelUserRead::lastRead) }
+        .stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    override val read: StateFlow<ChannelUserRead?> = _read
+
+    override val unreadCount: StateFlow<Int?> = _unreadCount
+
+    override val members: StateFlow<List<Member>> = _members
+        .combine(latestUsers) { membersMap, usersMap -> membersMap.values.updateUsers(usersMap) }
+        .map { it.sortedBy(Member::createdAt) }
+        .stateIn(scope, SharingStarted.Eagerly, emptyList())
+
+    override val membersCount: StateFlow<Int> = _membersCount
+
+    override val channelData: StateFlow<ChannelData> =
+        _channelData.filterNotNull().combine(latestUsers) { channelData, users ->
+            if (users.containsKey(channelData.createdBy.id)) {
+                channelData.copy(createdBy = users[channelData.createdBy.id] ?: channelData.createdBy)
+            } else {
+                channelData
+            }
+        }
+            .stateIn(scope, SharingStarted.Eagerly, ChannelData(type = channelType, channelId = channelId))
+
+    override val hidden: StateFlow<Boolean> = _hidden
+    override val muted: StateFlow<Boolean> = _muted
+    override val loading: StateFlow<Boolean> = _loading
+    override val loadingOlderMessages: StateFlow<Boolean> = _loadingOlderMessages
+    override val loadingNewerMessages: StateFlow<Boolean> = _loadingNewerMessages
+    override val endOfOlderMessages: StateFlow<Boolean> = _endOfOlderMessages
+    override val endOfNewerMessages: StateFlow<Boolean> = _endOfNewerMessages
 
     /** If we need to recover state when connection established again. */
-    override var recoveryNeeded: Boolean
+    override var recoveryNeeded: Boolean = false
+
+    override val insideSearch: StateFlow<Boolean> = _insideSearch
+
+    override fun toChannel(): Channel {
+        // recreate a channel object from the various observables.
+        val channelData = channelData.value
+
+        val messages = sortedMessages.value
+        val members = members.value
+        val watchers = watchers.value
+        val reads = _rawReads.value.values.toList()
+        val watcherCount = _watcherCount.value
+
+        val channel = channelData.toChannel(messages, members, reads, watchers, watcherCount)
+        channel.config = _channelConfig.value
+        channel.unreadCount = _unreadCount.value
+        channel.lastMessageAt =
+            lastMessageAt ?: messages.lastOrNull()?.let { it.createdAt ?: it.createdLocallyAt }
+        channel.hidden = _hidden.value
+
+        return channel
+    }
 
     /**
      * Sets the channel to be loading older messages.
      *
      * @param isLoading Boolean.
      */
-    fun setLoadingOlderMessages(isLoading: Boolean)
+    fun setLoadingOlderMessages(isLoading: Boolean) {
+        _loadingOlderMessages.value = isLoading
+    }
 
     /**
      * Sets the channel to be loading newer messages.
      *
      * @param isLoading Boolean.
      */
-    fun setLoadingNewerMessages(isLoading: Boolean)
+    fun setLoadingNewerMessages(isLoading: Boolean) {
+        _loadingNewerMessages.value = isLoading
+    }
 
     /**
      * Sets the watchers counter the this channel.
      *
      * @param count Int.
      */
-    fun setWatcherCount(count: Int)
+    fun setWatcherCount(count: Int) {
+        _watcherCount.value = count
+    }
 
     /**
      * Sets the read information for this channel.
      *
      * @param channelUserRead [ChannelUserRead]
      */
-    fun setRead(channelUserRead: ChannelUserRead?)
+    fun setRead(channelUserRead: ChannelUserRead?) {
+        _read.value = channelUserRead
+    }
 
     /** Sets the end for newer messages. */
-    fun setEndOfNewerMessages(isEnd: Boolean)
+    fun setEndOfNewerMessages(isEnd: Boolean) {
+        _endOfNewerMessages.value = isEnd
+    }
 
     /**
      * Sets the end for older messages.
      *
      * @param isEnd Boolean
      */
-    fun setEndOfOlderMessages(isEnd: Boolean)
+    fun setEndOfOlderMessages(isEnd: Boolean) {
+        _endOfOlderMessages.value = isEnd
+    }
 
     /**
      * Sets loadings.
      *
      * @param isLoading Boolean.
      */
-    fun setLoading(isLoading: Boolean)
+    fun setLoading(isLoading: Boolean) {
+        _loading.value = isLoading
+    }
 
     /**
      * Sets hidden.
      *
      * @param isHidden Boolean
      */
-    fun setHidden(isHidden: Boolean)
+    fun setHidden(isHidden: Boolean) {
+        _hidden.value = isHidden
+    }
 
     /**
      * Sets muted.
      *
      * @param isMuted Boolean.
      */
-    fun setMuted(isMuted: Boolean)
+    fun setMuted(isMuted: Boolean) {
+        _muted.value = isMuted
+    }
 
     /** Sets [ChannelData]. */
-    fun setChannelData(channelData: ChannelData)
+    fun setChannelData(channelData: ChannelData) {
+        _channelData.value = channelData
+    }
 
     /**
      * Sets replied message.
      *
      * @param repliedMessage [Message]
      */
-    fun setRepliedMessage(repliedMessage: Message?)
+    fun setRepliedMessage(repliedMessage: Message?) {
+        _repliedMessage.value = repliedMessage
+    }
 
     /**
      * Sets unread count.
      *
      * @param count Int.
      */
-    fun setUnreadCount(count: Int)
+    fun setUnreadCount(count: Int) {
+        _unreadCount.value = count
+    }
 
     /**
      *  Sets member count.
      *
      * @param count Int.
      */
-    fun setMembersCount(count: Int)
+    fun setMembersCount(count: Int) {
+        _membersCount.value = count
+    }
 
     /** Sets inside search. This must be set when a search is started in the channel and the
      * user is looks sort a older message using the search functionality.
      *
      * @param isInsideSearch Boolean.
      * */
-    fun setInsideSearch(isInsideSearch: Boolean)
+    fun setInsideSearch(isInsideSearch: Boolean) {
+        _insideSearch.value = isInsideSearch
+    }
 
     /**
      * Set channel config
      *
      * @param channelConfig [Config]
      */
-    fun setChannelConfig(channelConfig: Config)
+    fun setChannelConfig(channelConfig: Config) {
+        _channelConfig.value = channelConfig
+    }
 
     /**
      * Updates StateFlows related to typing updates.
      */
-    fun updateTypingEvents(eventsMap: Map<String, TypingStartEvent>, typingEvent: TypingEvent)
+    fun updateTypingEvents(eventsMap: Map<String, TypingStartEvent>, typingEvent: TypingEvent) {
+        _typingChatEvents.value = eventsMap
+        _typing.value = typingEvent
+    }
 }
+
+internal fun ChannelState.toMutableState(): ChannelMutableState = this as ChannelMutableState
