+++ stream-chat-android-state/src/test/java/io/getstream/chat/android/offline/channel/ChannelMarkReadHelperTests.kt
@@ -25,7 +25,7 @@ import io.getstream.chat.android.client.test.randomUser
 import io.getstream.chat.android.offline.plugin.logic.channel.internal.ChannelLogic
 import io.getstream.chat.android.offline.plugin.logic.internal.LogicRegistry
 import io.getstream.chat.android.offline.plugin.state.StateRegistry
-import io.getstream.chat.android.offline.plugin.state.channel.internal.ChannelMutableStateImpl
+import io.getstream.chat.android.offline.plugin.state.channel.internal.ChannelMutableState
 import io.getstream.chat.android.offline.utils.internal.ChannelMarkReadHelper
 import io.getstream.chat.android.test.TestCoroutineExtension
 import io.getstream.chat.android.test.randomCID
@@ -56,7 +56,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn false
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
         }
         val sut = Fixture()
@@ -74,7 +74,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn true
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(emptyList())
         }
@@ -94,7 +94,7 @@ internal class ChannelMarkReadHelperTests {
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
         val now = Date()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(listOf(randomMessage(createdAt = now)))
             on { it.lastMarkReadEvent } doReturn Date(now.time + 1000)
@@ -114,7 +114,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn true
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(listOf(randomMessage()))
         }
@@ -134,7 +134,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn true
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(listOf(randomMessage()))
         }
@@ -154,7 +154,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn true
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(listOf(randomMessage()))
         }
@@ -177,7 +177,7 @@ internal class ChannelMarkReadHelperTests {
             on { it.readEventsEnabled } doReturn true
         }
         val (channelType, channelId) = randomCID().cidToTypeAndId()
-        val channelMutableState = mock<ChannelMutableStateImpl> {
+        val channelMutableState = mock<ChannelMutableState> {
             on { it.channelConfig } doReturn MutableStateFlow(config)
             on { it.sortedMessages } doReturn MutableStateFlow(listOf(randomMessage()))
         }
@@ -200,7 +200,7 @@ internal class ChannelMarkReadHelperTests {
         private val state = mock<StateRegistry>()
         private val clientState = mock<ClientState>()
 
-        fun givenChannelState(channelType: String, channelId: String, channelState: ChannelMutableStateImpl) = apply {
+        fun givenChannelState(channelType: String, channelId: String, channelState: ChannelMutableState) = apply {
             whenever(state.channel(channelType = channelType, channelId = channelId)) doReturn channelState
         }
